pyproject.toml
[project]
name = "rdfcon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerberus>=1.3.7",
    "pyyaml>=6.0.3",
    "rdflib>=7.2.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.0.1",
    "ruff>=0.13.2",
]
